refactor: 절대경로 수정 로직분리
