+++ src/components/features/auth/signup/index.tsx
@@ -1,45 +1,22 @@
 import { useState } from 'react';
 import { Link, useNavigate } from 'react-router-dom';
+import dogImg from '@assets/dog_01.webp';
 import { Button, Input, message } from 'antd';
-import { createUserWithEmailAndPassword, getAuth } from 'firebase/auth';
 
-import { app } from './config';
+import { register } from '@/utils/authRegister';
+
 import * as S from './style';
-
-import dogImg from '/src/assets/dog_01.webp';
-
-interface SignUpProps {
-	email: string;
-	password: string;
-}
 
 const Signup = () => {
 	const [email, setEmail] = useState('');
 	const [password, setPassword] = useState('');
 	const navigate = useNavigate();
-	const [messageApi, contextHolder] = message.useMessage(); // message.useMessage()를 통해 contextHolder 변수 설정
-
-	async function register({ email, password }: SignUpProps) {
-		try {
-			const auth = getAuth(app);
-			const user = await createUserWithEmailAndPassword(auth, email, password);
-			console.log(user);
-
-			messageApi.info('Sign-up Successful!');
-			setTimeout(() => {
-				navigate('/login');
-			}, 2000);
-		} catch (error) {
-			messageApi.open({
-				type: 'error',
-				content: 'Please check the backup post again.',
-			});
-		}
-	}
+	const [, contextHolder] = message.useMessage();
 
 	const onSubmit = (e: React.FormEvent<HTMLFormElement>) => {
 		e.preventDefault();
-		register({ email, password });
+		// 비즈니스로직 분리
+		register({ email, password }, navigate);
 	};
 
 	return (
@@ -69,7 +46,7 @@ const Signup = () => {
 					<Link to='/login'>Already have an account? Login!</Link>
 				</S.SignUpFormCont>
 			</S.SignUpWrap>
-			{contextHolder} {/* contextHolder를 여기서 출력 */}
+			{contextHolder}
 		</S.SignUpContainer>
 	);
 };
