+++ src/interfaces/auth.ts
@@ -1,0 +1,4 @@
+export interface SignUpProps {
+	email: string;
+	password: string;
+}
